pull out inputs
